+++ src/App.js
@@ -1,23 +1,23 @@
 import React from 'react';
 import './App.css';
 import Header from 'component/Header.js';
-import Profile from 'component/Profile/Profile.js';
-import Map from 'component/Map/Map.js';
-import Login from 'component/Login/Login';
-import SignUp from 'component/SignUp/SignUp';
+import Profile from 'component/pages/Profile/Profile.js';
+import Map from 'component/pages/Map/Map.js';
+import Login from 'component/pages/Login/Login';
+import SignUp from 'component/pages/SignUp/SignUp';
 
 class App extends React.Component {
   constructor(){
     super();
     this.state = {
-      type: "SIGNUP",
+      page: "SIGNUP",
     }
 
     this.changePage = this.changePage.bind(this);
   }
 
-  changePage(type) {
-    this.setState({type});
+  changePage(page) {
+    this.setState({page});
   } 
 
   render() {
@@ -28,9 +28,9 @@ class App extends React.Component {
         {
           PROFILE: <Profile />,
           MAP: <Map />,
-          LOGIN: <Login onSubmit={this.changePage} />,
-          SIGNUP: <SignUp onSubmit={this.changePage} />
-        }[this.state.type]
+          LOGIN: <Login changePage={this.changePage} />,
+          SIGNUP: <SignUp changePage={this.changePage} />
+        }[this.state.page]
       }
     </div>
     );
