+++ src/component/Header.js
@@ -3,13 +3,6 @@ import React from 'react';
 import logo from 'logo-taxi.svg';
 
 class Header extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-
-        };
-    }
-
     render() {
         return (
             <header>
