+++ src/component/pages/SignUp/SignUp.js
@@ -3,21 +3,33 @@ import PropTypes from 'prop-types';
 import { Container, TextField } from '@material-ui/core'; 
 
 class SignUp extends React.Component {
+    onSubmit = (e) => {
+        e.preventDefault();
+        this.props.changePage("MAP");
+    }
+    
     render() {
         return (
             <div>
-                Уже зарегистрированы? <a href="#" onClick={() => {this.props.onSubmit("LOGIN")}}>Войти</a>
-                <form onSubmit={(e) => {e.preventDefault(); this.props.onSubmit("MAP")}}>
-                    <label> Адрес электронной почты
+                Уже зарегистрированы?
+                <a href="#" onClick={() => {this.props.changePage("LOGIN")}}>
+                    Войти
+                </a>
+                <form onSubmit={this.onSubmit}>
+                    <label>
+                        Адрес электронной почты
                         <input name="email" type="email" />
                     </label>
-                    <label> Имя
+                    <label>
+                        Имя
                         <input name="name" type="text" />
                     </label>
-                    <label> Фамилия
+                    <label>
+                        Фамилия
                         <input name="surname" type="text" />
                     </label>
-                    <label> Пароль
+                    <label>
+                        Пароль
                         <input name="password" type="password" />
                     </label>
                     <button type="submit">Войти</button>
@@ -28,7 +40,7 @@ class SignUp extends React.Component {
 }
 
 SignUp.propTypes = {
-    onSubmit: PropTypes.func
+    changePage: PropTypes.func
 };
 
 export default SignUp;
